[23] Move error handling in a concern
