+++ app/controllers/todos_controller.rb
@@ -1,9 +1,7 @@
 # frozen_string_literal: true
 
 class TodosController < ApplicationController
-  rescue_from ActiveRecord::RecordInvalid, with: :record_invalid_handler
-  rescue_from ActiveRecord::RecordNotFound, with: :record_not_found_handler
-  rescue_from ActionController::ParameterMissing, with: :parameter_missing_handler
+  include ErrorHandler
 
   def index
     todos = Todo.all
@@ -52,29 +50,5 @@ class TodosController < ApplicationController
 
   def update_todo_params
     params.permit(:title, :order, :completed)
-  end
-
-  def record_invalid_handler(exception)
-    errors = exception.record.errors.map do |error|
-      error_message(error.full_message, nil, Rack::Utils.status_code(:unprocessable_entity))
-    end
-
-    render json: { errors: }, status: :unprocessable_entity
-  end
-
-  def record_not_found_handler(exception)
-    error = error_message(exception.message, nil, Rack::Utils.status_code(:not_found))
-
-    render json: { errors: [error] }, status: :not_found
-  end
-
-  def parameter_missing_handler(exception)
-    error = error_message(exception.message, nil, Rack::Utils.status_code(:unprocessable_entity))
-
-    render json: { errors: [error] }, status: :unprocessable_entity
-  end
-
-  def error_message(title, detail, status)
-    { title:, detail:, status: }
   end
 end
